Prompt for configuration if files missing
This commit introduces a big chunk of new functionalities. If, at
bootstrap, the configuration files are missing (as is the default case
when creating a new mailbundle), the user is now prompted to insert the
needed data.
To ease the configuration burden, a series of heuristics are introduced
to find the IMAP and SMTP connection endpoint for a new account.